Fix various build errors in Xwt.Mac project (that were not fixed after last merge with main XWT branch)

+++ Xwt.XamMac/Xwt.Mac/ImageHandler.cs
@@ -185,7 +185,7 @@ namespace Xwt.Mac
 					im.Size = new CGSize ((nfloat)width, (nfloat)height);
 					return im;
 				}
-				bitmap.Size = new CGSize(width, height);
+				bitmap.Size = new CGSize((float)width, (float)height);
 				return handle;
 			}
 		}
